+++ android/src/main/java/com/applovin/reactnative/AppLovinMAXNativeAdView.java
@@ -54,6 +54,8 @@ public class AppLovinMAXNativeAdView
     private String              customData;
     @Nullable
     private Map<String, Object> extraParameters;
+    @Nullable
+    private Map<String, Object> localExtraParameters;
 
     // TODO: Allow publisher to select which views are clickable and which isn't via prop
     private final List<View> clickableViews = new ArrayList<>();
@@ -103,6 +105,14 @@ public class AppLovinMAXNativeAdView
         }
     }
 
+    public void setLocalExtraParameters(@Nullable final ReadableMap readableMap)
+    {
+        if ( readableMap != null )
+        {
+            localExtraParameters = readableMap.toHashMap();
+        }
+    }
+
     public void loadAd()
     {
         if ( isLoading.compareAndSet( false, true ) )
@@ -124,6 +134,14 @@ public class AppLovinMAXNativeAdView
                 for ( Map.Entry<String, Object> entry : extraParameters.entrySet() )
                 {
                     adLoader.setExtraParameter( entry.getKey(), (String) entry.getValue() );
+                }
+            }
+
+            if ( localExtraParameters != null )
+            {
+                for ( Map.Entry<String, Object> entry : localExtraParameters.entrySet() )
+                {
+                    adLoader.setLocalExtraParameter( entry.getKey(), (String) entry.getValue() );
                 }
             }
 
@@ -366,7 +384,7 @@ public class AppLovinMAXNativeAdView
         if ( !Float.isNaN( aspectRatio ) )
         {
             // The aspect ratio can be 0.0f when it is not provided by the network.
-            if ( Math.signum( aspectRatio )  == 0 )
+            if ( Math.signum( aspectRatio ) == 0 )
             {
                 nativeAdInfo.putDouble( "mediaContentAspectRatio", 1.0 );
             }