+++ ios/AppLovinMAXAdViewManager.m
@@ -19,6 +19,8 @@ RCT_EXPORT_VIEW_PROPERTY(placement, NSString)
 RCT_EXPORT_VIEW_PROPERTY(customData, NSString)
 RCT_EXPORT_VIEW_PROPERTY(adaptiveBannerEnabled, BOOL)
 RCT_EXPORT_VIEW_PROPERTY(autoRefresh, BOOL)
+RCT_EXPORT_VIEW_PROPERTY(extraParameters, NSDictionary)
+RCT_EXPORT_VIEW_PROPERTY(localExtraParameters, NSDictionary)
 
 RCT_EXPORT_VIEW_PROPERTY(onAdLoadedEvent, RCTDirectEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(onAdLoadFailedEvent, RCTDirectEventBlock)