+++ src/NativeAdView.js
@@ -40,6 +40,8 @@ const AppLovinMAXNativeAdView = requireNativeComponent('AppLovinMAXNativeAdView'
 // 3. update of the nativeAd context by onNativeAdLoaded, which renders the ad components with nativeAd
 const NativeAdView = forwardRef((props, ref) => {
 
+  const {extraParameters, localExtraParameters, ...otherProps} = props;
+
   // context from NativeAdViewProvider
   const {nativeAd, nativeAdView, setNativeAd, setNativeAdView} = useContext(NativeAdViewContext);
 
@@ -85,14 +87,29 @@ const NativeAdView = forwardRef((props, ref) => {
     if (props.onAdRevenuePaid) props.onAdRevenuePaid(event.nativeEvent);
   };
 
+  const checkExtraParameters = (name, params) => {
+    if (params) {
+      for (const key in params) {
+        const value = params[key];
+        if ((value != null) && (value != undefined) && (typeof value !== 'string')) {
+          console.warn(name + " in NativeAdView supports only string values: " + key);
+          delete params[key];
+        }
+      }
+    }
+    return params;
+  };
+
   return (
     <AppLovinMAXNativeAdView
       ref={saveElement}
+      extraParameters={checkExtraParameters('extraParameters', extraParameters)}
+      localExtraParameters={checkExtraParameters('localExtraParameters', localExtraParameters)}
       onAdLoadedEvent={onAdLoadedEvent}
       onAdLoadFailedEvent={onAdLoadFailedEvent}
       onAdClickedEvent={onAdClickedEvent}
       onAdRevenuePaidEvent={onAdRevenuePaidEvent}
-      {...props}
+      {...otherProps}
     >
       {props.children}
     </AppLovinMAXNativeAdView>
@@ -119,6 +136,11 @@ NativeAdView.propTypes = {
    * A dictionary value representing the extra parameters to set a list of key-value string pairs.
    */
   extraParameters: PropTypes.object,
+
+  /**
+   * A dictionary value representing the local extra parameters to set a list of key-value string pairs.
+   */
+  localExtraParameters: PropTypes.object,
 
   /**
    * A callback fuction to be fired when a new ad has been loaded.